Refactor core utilities and network components into `core.common` package
- Relocate `Constants`, `Currency` utilities, `CsDispatchers`, and coroutine DI modules to the `ru.resodostudios.cashsense.core.common` package.
- Update all references to `getUsdCurrency`, `getDefaultCurrency`, `CsDispatchers`, and `ApplicationScope` across the project.
- Remove unused dependencies and simplify the `build.gradle.kts` for `feature:transaction:csvimport:impl`.
- Reorganize transaction-related module includes in `settings.gradle.kts`.
