+++ core/common/src/main/kotlin/ru/resodostudios/cashsense/core/common/CsDispatchers.kt
@@ -1,4 +1,4 @@
-package ru.resodostudios.cashsense.core.network
+package ru.resodostudios.cashsense.core.common
 
 import javax.inject.Qualifier
 import kotlin.annotation.AnnotationRetention.RUNTIME
