+++ core/common/src/main/kotlin/ru/resodostudios/cashsense/core/common/di/DispatchersModule.kt
@@ -1,4 +1,4 @@
-package ru.resodostudios.cashsense.core.network.di
+package ru.resodostudios.cashsense.core.common.di
 
 import dagger.Module
 import dagger.Provides
@@ -6,9 +6,9 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
-import ru.resodostudios.cashsense.core.network.CsDispatchers.Default
-import ru.resodostudios.cashsense.core.network.CsDispatchers.IO
-import ru.resodostudios.cashsense.core.network.Dispatcher
+import ru.resodostudios.cashsense.core.common.CsDispatchers.Default
+import ru.resodostudios.cashsense.core.common.CsDispatchers.IO
+import ru.resodostudios.cashsense.core.common.Dispatcher
 
 @Module
 @InstallIn(SingletonComponent::class)
