+++ core/datastore/src/main/kotlin/ru/resodostudios/cashsense/core/datastore/UserPreferencesSerializer.kt
@@ -3,7 +3,7 @@ package ru.resodostudios.cashsense.core.datastore
 import androidx.datastore.core.CorruptionException
 import androidx.datastore.core.Serializer
 import com.google.protobuf.InvalidProtocolBufferException
-import ru.resodostudios.cashsense.core.util.getDefaultCurrency
+import ru.resodostudios.cashsense.core.common.getDefaultCurrency
 import java.io.InputStream
 import java.io.OutputStream
 import javax.inject.Inject
