+++ core/datastore/src/main/kotlin/ru/resodostudios/cashsense/core/datastore/di/DataStoreModule.kt
@@ -11,11 +11,11 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
+import ru.resodostudios.cashsense.core.common.CsDispatchers.IO
+import ru.resodostudios.cashsense.core.common.Dispatcher
+import ru.resodostudios.cashsense.core.common.di.ApplicationScope
 import ru.resodostudios.cashsense.core.datastore.UserPreferences
 import ru.resodostudios.cashsense.core.datastore.UserPreferencesSerializer
-import ru.resodostudios.cashsense.core.network.CsDispatchers.IO
-import ru.resodostudios.cashsense.core.network.Dispatcher
-import ru.resodostudios.cashsense.core.network.di.ApplicationScope
 import javax.inject.Singleton
 
 @Module
