+++ core/notifications/src/main/kotlin/ru/resodostudios/cashsense/core/notifications/SystemTrayNotifier.kt
@@ -17,11 +17,11 @@ import androidx.core.net.toUri
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toJavaZoneId
+import ru.resodostudios.cashsense.core.common.Constants.DEEPLINK_PATH_BASE
+import ru.resodostudios.cashsense.core.common.Constants.DEEPLINK_TAG_SUBSCRIPTIONS
+import ru.resodostudios.cashsense.core.common.Constants.TARGET_ACTIVITY_NAME
 import ru.resodostudios.cashsense.core.model.data.Subscription
 import ru.resodostudios.cashsense.core.ui.util.formatAmount
-import ru.resodostudios.cashsense.core.util.Constants.DEEPLINK_PATH_BASE
-import ru.resodostudios.cashsense.core.util.Constants.DEEPLINK_TAG_SUBSCRIPTIONS
-import ru.resodostudios.cashsense.core.util.Constants.TARGET_ACTIVITY_NAME
 import java.time.format.DateTimeFormatter
 import java.time.format.FormatStyle
 import java.util.Locale
