+++ feature/subscription/list/impl/src/main/kotlin/ru/resodostudios/cashsense/feature/subscription/list/impl/SubscriptionPreviewParameterProvider.kt
@@ -1,10 +1,10 @@
 package ru.resodostudios.cashsense.feature.subscription.list.impl
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import ru.resodostudios.cashsense.core.common.getUsdCurrency
 import ru.resodostudios.cashsense.core.model.data.RepeatingIntervalType.MONTHLY
 import ru.resodostudios.cashsense.core.model.data.RepeatingIntervalType.NONE
 import ru.resodostudios.cashsense.core.model.data.Subscription
-import ru.resodostudios.cashsense.core.util.getUsdCurrency
 import java.math.BigDecimal
 import kotlin.time.Instant
