+++ feature/transaction/csvimport/impl/build.gradle.kts
@@ -10,16 +10,8 @@ android {
 dependencies {
     implementation(projects.feature.transaction.csvimport.api)
 
-    implementation(projects.core.common)
     implementation(projects.core.domain)
-    implementation(projects.core.model)
-    implementation(projects.core.ui)
-    implementation(projects.core.designsystem)
-    implementation(projects.core.locales)
 
-    implementation(libs.kotlin.csv)
-    implementation(libs.androidx.hilt.lifecycle.viewmodel.compose)
-    implementation(libs.androidx.lifecycle.runtimeCompose)
-    implementation(libs.androidx.lifecycle.viewModelCompose)
     implementation(libs.androidx.navigation3.ui)
+    implementation(libs.kotlin.csv)
 }
