+++ feature/transaction/detail/impl/src/main/kotlin/ru/resodostudios/cashsense/feature/transaction/detail/impl/TransactionViewModel.kt
@@ -14,9 +14,9 @@ import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
+import ru.resodostudios.cashsense.core.common.di.ApplicationScope
 import ru.resodostudios.cashsense.core.data.repository.TransactionsRepository
 import ru.resodostudios.cashsense.core.model.data.Transaction
-import ru.resodostudios.cashsense.core.network.di.ApplicationScope
 import ru.resodostudios.cashsense.feature.transaction.detail.api.TransactionNavKey
 import kotlin.time.Duration.Companion.seconds
