+++ feature/transfer/impl/src/main/kotlin/ru/resodostudios/cashsense/feature/transfer/impl/TransferDialogViewModel.kt
@@ -13,12 +13,12 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import ru.resodostudios.cashsense.core.common.di.ApplicationScope
+import ru.resodostudios.cashsense.core.common.getUsdCurrency
 import ru.resodostudios.cashsense.core.data.repository.TransactionsRepository
 import ru.resodostudios.cashsense.core.data.repository.WalletsRepository
 import ru.resodostudios.cashsense.core.model.data.Transaction
 import ru.resodostudios.cashsense.core.model.data.Transfer
-import ru.resodostudios.cashsense.core.network.di.ApplicationScope
-import ru.resodostudios.cashsense.core.util.getUsdCurrency
 import ru.resodostudios.cashsense.feature.transfer.dialog.api.TransferDialogNavKey
 import java.math.BigDecimal
 import java.math.RoundingMode
