+++ feature/wallet/widget/src/main/kotlin/ru/resodostudios/cashsense/wallet/widget/ui/WalletWidget.kt
@@ -36,13 +36,13 @@ import androidx.glance.layout.padding
 import androidx.glance.text.Text
 import dagger.hilt.android.EntryPointAccessors
 import kotlinx.coroutines.flow.first
+import ru.resodostudios.cashsense.core.common.Constants.DEEPLINK_PATH_BASE
+import ru.resodostudios.cashsense.core.common.Constants.DEEPLINK_TAG_HOME
+import ru.resodostudios.cashsense.core.common.Constants.DEEPLINK_TAG_TRANSACTION
+import ru.resodostudios.cashsense.core.common.Constants.DEEPLINK_TAG_WALLET
+import ru.resodostudios.cashsense.core.common.Constants.TARGET_ACTIVITY_NAME
 import ru.resodostudios.cashsense.core.model.data.ExtendedUserWallet
 import ru.resodostudios.cashsense.core.ui.util.formatAmount
-import ru.resodostudios.cashsense.core.util.Constants.DEEPLINK_PATH_BASE
-import ru.resodostudios.cashsense.core.util.Constants.DEEPLINK_TAG_HOME
-import ru.resodostudios.cashsense.core.util.Constants.DEEPLINK_TAG_TRANSACTION
-import ru.resodostudios.cashsense.core.util.Constants.DEEPLINK_TAG_WALLET
-import ru.resodostudios.cashsense.core.util.Constants.TARGET_ACTIVITY_NAME
 import ru.resodostudios.cashsense.feature.wallet.widget.R
 import ru.resodostudios.cashsense.wallet.widget.WalletWidgetEntryPoint
 import ru.resodostudios.cashsense.wallet.widget.ui.theme.CsGlanceTheme
