+++ settings.gradle.kts
@@ -61,10 +61,10 @@ include(":feature:subscription:dialog:api")
 include(":feature:subscription:dialog:impl")
 include(":feature:subscription:list:api")
 include(":feature:subscription:list:impl")
-include(":feature:transaction:detail:api")
-include(":feature:transaction:detail:impl")
 include(":feature:transaction:csvimport:api")
 include(":feature:transaction:csvimport:impl")
+include(":feature:transaction:detail:api")
+include(":feature:transaction:detail:impl")
 include(":feature:transaction:editor:api")
 include(":feature:transaction:editor:impl")
 include(":feature:transaction:overview:api")
